@@ -55,7 +55,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.7</java.version>
+        <java.version>1.8</java.version>
         <version.unified>1.5.5</version.unified>
         <cluster.secured>false</cluster.secured>
         <cluster.type>mini</cluster.type>
@@ -137,7 +137,7 @@
         <dependency>
             <groupId>org.apache.hadoop</groupId>
             <artifactId>hadoop-common</artifactId>
-            <version>3.1.2</version>
+            <version>2.7.7</version>
             <scope>compile</scope>
         </dependency>
 
