@@ -20,9 +20,10 @@ import com.kakao.hbase.common.Args;
 import com.kakao.hbase.specific.CommandAdapter;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.ClusterStatus;
-import org.apache.hadoop.hbase.HRegionInfo;
 import org.apache.hadoop.hbase.ServerName;
-import org.apache.hadoop.hbase.client.HBaseAdmin;
+import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.Admin;
+import org.apache.hadoop.hbase.client.RegionInfo;
 import org.apache.hadoop.hbase.master.RegionPlan;
 
 import java.io.IOException;
@@ -32,14 +33,14 @@ import java.util.*;
 enum BalanceRule {
     RR {
         @Override
-        List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) throws IOException {
+        List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) throws IOException {
             List<ServerName> serverNames = getServerNames(admin);
-            Map<HRegionInfo, ServerName> regionLocations = Balance.getRegionAssignmentMap(admin, tableNameSet);
+            Map<RegionInfo, ServerName> regionLocations = Balance.getRegionAssignmentMap(admin, tableNameSet);
 
             int i = 0;
             List<RegionPlan> regionPlanList = new ArrayList<>();
-            for (String tableName : tableNameSet) {
-                for (HRegionInfo hRegionInfo : admin.getTableRegions(tableName.getBytes())) {
+            for (TableName tableName : tableNameSet) {
+                for (RegionInfo hRegionInfo : admin.getTableRegions(tableName)) {
                     ServerName source = regionLocations.get(hRegionInfo);
                     ServerName dest = serverNames.get((i++) % serverNames.size());
                     if (source == null || !source.equals(dest))
@@ -51,14 +52,14 @@ enum BalanceRule {
     },
     RD {
         @Override
-        List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) throws IOException {
+        List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) throws IOException {
             List<ServerName> serverNames = getServerNames(admin);
             List<ServerName> serverNamesToMove = new ArrayList<>(serverNames);
-            Map<HRegionInfo, ServerName> regionLocations = Balance.getRegionAssignmentMap(admin, tableNameSet);
+            Map<RegionInfo, ServerName> regionLocations = Balance.getRegionAssignmentMap(admin, tableNameSet);
 
             List<RegionPlan> regionPlanList = new ArrayList<>();
-            for (String tableName : tableNameSet) {
-                for (HRegionInfo hRegionInfo : admin.getTableRegions(tableName.getBytes())) {
+            for (TableName tableName : tableNameSet) {
+                for (RegionInfo hRegionInfo : admin.getTableRegions(tableName)) {
                     if (serverNamesToMove.size() == 0)
                         serverNamesToMove = new ArrayList<>(serverNames);
 
@@ -73,16 +74,16 @@ enum BalanceRule {
     },
     ST {
         @Override
-        List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) throws IOException {
+        List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) throws IOException {
             return makeStochasticPlan(admin, tableNameSet, BalanceFactor.parseArg(args));
         }
     },
     ST2 {
         @Override
-        List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) throws IOException {
+        List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) throws IOException {
             List<RegionPlan> regionPlanList = new ArrayList<>();
-            for (final String tableName : tableNameSet) {
-                HashSet<String> singleTableName = new HashSet<>();
+            for (TableName tableName : tableNameSet) {
+                Set<TableName> singleTableName = new HashSet<>();
                 singleTableName.add(tableName);
                 List<RegionPlan> regionPlans = makeStochasticPlan(admin, singleTableName, BalanceFactor.parseArg(args));
                 regionPlanList.addAll(regionPlans);
@@ -93,19 +94,19 @@ enum BalanceRule {
     },
     DEFAULT {
         @Override
-        List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) {
+        List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) {
             throw new IllegalStateException("do not call me");
         }
     };
 
     private static List<RegionPlan> makeStochasticPlan(
-            HBaseAdmin admin, Set<String> tableNameSet, BalanceFactor balanceFactor) throws IOException
+            Admin admin, Set<TableName> tableNameSet, BalanceFactor balanceFactor) throws IOException
     {
-        Map<ServerName, List<HRegionInfo>> clusterState = CommandAdapter.initializeRegionMap(admin);
+        Map<ServerName, List<RegionInfo>> clusterState = CommandAdapter.initializeRegionMap(admin);
 
-        Map<HRegionInfo, ServerName> regionAssignmentMap = Balance.getRegionAssignmentMap(admin, tableNameSet);
-        for (Map.Entry<HRegionInfo, ServerName> entry : regionAssignmentMap.entrySet()) {
-            List<HRegionInfo> hRegionInfos = clusterState.get(entry.getValue());
+        Map<RegionInfo, ServerName> regionAssignmentMap = Balance.getRegionAssignmentMap(admin, tableNameSet);
+        for (Map.Entry<RegionInfo, ServerName> entry : regionAssignmentMap.entrySet()) {
+            List<RegionInfo> hRegionInfos = clusterState.get(entry.getValue());
             if (hRegionInfos != null) {
                 hRegionInfos.add(entry.getKey());
             }
@@ -118,12 +119,12 @@ enum BalanceRule {
         return CommandAdapter.makePlan(admin, clusterState, conf);
     }
 
-    static List<ServerName> getServerNames(HBaseAdmin admin) throws IOException {
+    static List<ServerName> getServerNames(Admin admin) throws IOException {
         ClusterStatus clusterStatus = admin.getClusterStatus();
         List<ServerName> serverNameList = new ArrayList<>(clusterStatus.getServers());
         Collections.sort(serverNameList);
         return serverNameList;
     }
 
-    abstract List<RegionPlan> makePlan(HBaseAdmin admin, Set<String> tableNameSet, Args args) throws IOException;
+    abstract List<RegionPlan> makePlan(Admin admin, Set<TableName> tableNameSet, Args args) throws IOException;
 }
