@@ -20,8 +20,9 @@ import com.kakao.hbase.ManagerArgs;
 import com.kakao.hbase.TestBase;
 import com.kakao.hbase.common.Args;
 import com.kakao.hbase.common.Constant;
-import org.apache.hadoop.hbase.HRegionInfo;
 import org.apache.hadoop.hbase.ServerName;
+import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.RegionInfo;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -38,13 +39,13 @@ public class AssignRestoreTest extends TestBase {
     @Test
     public void testRestoreTable() throws Exception {
         ArrayList<ServerName> serverNameList;
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
         ServerName rs1, rs2;
         String[] argsParam;
         Args args;
         Assign command;
 
-        String tableName2 = createAdditionalTable(tableName + 2);
+        TableName tableName2 = createAdditionalTable(tableName + "2");
         splitTable("a".getBytes());
         splitTable("b".getBytes());
         regionInfoList = getRegionInfoList(tableName);
@@ -56,7 +57,7 @@ public class AssignRestoreTest extends TestBase {
         rs2 = serverNameList.get(1);
 
         // move all regions to rs1
-        for (HRegionInfo hRegionInfo : getRegionInfoList(tableName))
+        for (RegionInfo hRegionInfo : getRegionInfoList(tableName))
             move(hRegionInfo, rs1);
         move(getRegionInfoList(tableName2).get(0), rs1);
         assertEquals(3, getRegionInfoList(rs1, tableName).size());
@@ -66,7 +67,7 @@ public class AssignRestoreTest extends TestBase {
         String timestamp1 = Constant.DATE_FORMAT_ARGS.format(System.currentTimeMillis());
 
         // move all regions to rs2 for tableName only
-        for (HRegionInfo hRegionInfo : getRegionInfoList(tableName))
+        for (RegionInfo hRegionInfo : getRegionInfoList(tableName))
             move(hRegionInfo, rs2);
         assertEquals(0, getRegionInfoList(rs1, tableName).size());
         assertEquals(3, getRegionInfoList(rs2, tableName).size());
@@ -84,10 +85,10 @@ public class AssignRestoreTest extends TestBase {
 
         boolean balancerRunning = false;
         try {
-            balancerRunning = admin.setBalancerRunning(false, true);
+            balancerRunning = admin.balancerSwitch(false, true);
 
             // restore table with timestamp1
-            argsParam = new String[]{"zookeeper", "restoRe", "tAble", tableName, timestamp1, "--force-proceed"};
+            argsParam = new String[]{"zookeeper", "restoRe", "tAble", tableName.getNameAsString(), timestamp1, "--force-proceed"};
             args = new ManagerArgs(argsParam);
             command = new Assign(admin, args);
             command.run();
@@ -98,7 +99,7 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(2, AssignAction.getProcessedCount());
 
             // restore table with timestamp2
-            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName, timestamp2, "--force-proceed"};
+            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName.getNameAsString(), timestamp2, "--force-proceed"};
             args = new ManagerArgs(argsParam);
             command = new Assign(admin, args);
             command.run();
@@ -109,7 +110,7 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(3, AssignAction.getProcessedCount());
 
             // restore table with timestamp3
-            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName, timestamp3, "--force-proceed"};
+            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName.getNameAsString(), timestamp3, "--force-proceed"};
             args = new ManagerArgs(argsParam);
             command = new Assign(admin, args);
             command.run();
@@ -120,10 +121,11 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(1, AssignAction.getProcessedCount());
 
             // restore table with timestamp1
-            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName, timestamp1, "--force-proceed"};
+            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName.getNameAsString(), timestamp1, "--force-proceed"};
             args = new ManagerArgs(argsParam);
             command = new Assign(admin, args);
             command.run();
+            // fixme
             assertEquals(3, getRegionInfoList(rs1, tableName).size());
             assertEquals(0, getRegionInfoList(rs2, tableName).size());
             assertEquals(1, getRegionInfoList(rs1, tableName2).size());
@@ -131,13 +133,13 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(2, AssignAction.getProcessedCount());
         } finally {
             if (balancerRunning)
-                admin.setBalancerRunning(true, true);
+                admin.balancerSwitch(true, true);
         }
     }
 
     @Test
     public void testRestoreTableException() throws Exception {
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
         String[] argsParam;
         Args args;
         Assign command;
@@ -150,10 +152,10 @@ public class AssignRestoreTest extends TestBase {
 
         boolean balancerRunning = false;
         try {
-            balancerRunning = admin.setBalancerRunning(false, true);
+            balancerRunning = admin.balancerSwitch(false, true);
 
             // restore table with timestamp1
-            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName, timestamp1, "--force-proceed"};
+            argsParam = new String[]{"zookeeper", "restoRe", "table", tableName.getNameAsString(), timestamp1, "--force-proceed"};
             args = new ManagerArgs(argsParam);
             command = new Assign(admin, args);
             command.run();
@@ -161,20 +163,20 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(0, AssignAction.getProcessedCount());
         } finally {
             if (balancerRunning)
-                admin.setBalancerRunning(true, true);
+                admin.balancerSwitch(true, true);
         }
     }
 
     @Test
     public void testRestoreTableRegex() throws Exception {
         ArrayList<ServerName> serverNameList;
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
         ServerName rs1, rs2;
         String[] argsParam;
         Args args;
         Assign command;
 
-        String tableName2 = createAdditionalTable(tableName + 2);
+        TableName tableName2 = createAdditionalTable(tableName + "2");
         splitTable("a".getBytes());
         regionInfoList = getRegionInfoList(tableName);
         assertEquals(2, regionInfoList.size());
@@ -205,7 +207,7 @@ public class AssignRestoreTest extends TestBase {
 
         boolean balancerRunning = false;
         try {
-            balancerRunning = admin.setBalancerRunning(false, true);
+            balancerRunning = admin.balancerSwitch(false, true);
             String regex = tableName + ".*";
 
             // restore table with timestamp1
@@ -231,20 +233,20 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(3, AssignAction.getProcessedCount());
         } finally {
             if (balancerRunning)
-                admin.setBalancerRunning(true, true);
+                admin.balancerSwitch(true, true);
         }
     }
 
     @Test
     public void testRestoreSingleRs() throws Exception {
         ArrayList<ServerName> serverNameList;
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
         ServerName rs1, rs2;
         String[] argsParam;
         Args args;
         Assign command;
 
-        String tableName2 = createAdditionalTable(tableName + 2);
+        TableName tableName2 = createAdditionalTable(tableName + "2");
         splitTable("a".getBytes());
         regionInfoList = getRegionInfoList(tableName);
         assertEquals(2, regionInfoList.size());
@@ -275,7 +277,7 @@ public class AssignRestoreTest extends TestBase {
 
         boolean balancerRunning = false;
         try {
-            balancerRunning = admin.setBalancerRunning(false, true);
+            balancerRunning = admin.balancerSwitch(false, true);
             String regex = rs1.getHostname() + "," + rs1.getPort() + ".*";
 
             // restore table with timestamp1
@@ -301,20 +303,20 @@ public class AssignRestoreTest extends TestBase {
             assertTrue(AssignAction.getProcessedCount() >= 2);
         } finally {
             if (balancerRunning)
-                admin.setBalancerRunning(true, true);
+                admin.balancerSwitch(true, true);
         }
     }
 
     @Test
     public void testRestoreRsRegex() throws Exception {
         ArrayList<ServerName> serverNameList;
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
         ServerName rs1, rs2;
         String[] argsParam;
         Args args;
         Assign command;
 
-        String tableName2 = createAdditionalTable(tableName + 2);
+        TableName tableName2 = createAdditionalTable(tableName + "2");
         splitTable("a".getBytes());
         regionInfoList = getRegionInfoList(tableName);
         assertEquals(2, regionInfoList.size());
@@ -343,9 +345,13 @@ public class AssignRestoreTest extends TestBase {
         assertEquals(1, getRegionInfoList(rs2, tableName2).size());
         String timestamp2 = Constant.DATE_FORMAT_ARGS.format(System.currentTimeMillis());
 
+        System.out.println("111");
+        System.out.println(getRegionInfoList(tableName));
+        System.out.println(getRegionInfoList(tableName2));
+
         boolean balancerRunning = false;
         try {
-            balancerRunning = admin.setBalancerRunning(false, true);
+            balancerRunning = admin.balancerSwitch(false, true);
             String regex = rs1.getHostname() + ".*";
 
             // restore table with timestamp1
@@ -369,7 +375,7 @@ public class AssignRestoreTest extends TestBase {
             assertEquals(1, getRegionInfoList(rs2, tableName2).size());
         } finally {
             if (balancerRunning)
-                admin.setBalancerRunning(true, true);
+                admin.balancerSwitch(true, true);
         }
     }
 }
