@@ -19,7 +19,7 @@ package com.kakao.hbase.manager.command;
 import com.kakao.hbase.ManagerArgs;
 import com.kakao.hbase.common.Args;
 import com.kakao.hbase.common.Constant;
-import org.apache.hadoop.hbase.HRegionInfo;
+import org.apache.hadoop.hbase.client.RegionInfo;
 import org.junit.Test;
 
 import java.util.List;
@@ -36,10 +36,10 @@ public class MergeEmptyFastTest2 extends MergeTestBase {
     public void testMergeEmptyFast4() throws Exception {
         makeTestData4();
 
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
 
         // merge
-        String[] argsParam = {"zookeeper", tableName, "empty-FAST", "--force-proceed", "--test"};
+        String[] argsParam = {"zookeeper", tableName.getNameAsString(), "empty-FAST", "--force-proceed", "--test"};
         Args args = new ManagerArgs(argsParam);
         Merge command = new Merge(admin, args);
         command.run();
@@ -56,10 +56,10 @@ public class MergeEmptyFastTest2 extends MergeTestBase {
     public void testMergeEmptyFast5() throws Exception {
         makeTestData5();
 
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
 
         // merge
-        String[] argsParam = {"zookeeper", tableName, "empty-FAST", "--force-proceed", "--test"};
+        String[] argsParam = {"zookeeper", tableName.getNameAsString(), "empty-FAST", "--force-proceed", "--test"};
         Args args = new ManagerArgs(argsParam);
         Merge command = new Merge(admin, args);
         command.run();
@@ -74,10 +74,10 @@ public class MergeEmptyFastTest2 extends MergeTestBase {
     public void testMergeEmptyFast6() throws Exception {
         makeTestData6();
 
-        List<HRegionInfo> regionInfoList;
+        List<RegionInfo> regionInfoList;
 
         // merge
-        String[] argsParam = {"zookeeper", tableName, "empty-FAST", "--force-proceed", "--test"
+        String[] argsParam = {"zookeeper", tableName.getNameAsString(), "empty-FAST", "--force-proceed", "--test"
                 , "--verbose", "--debug"};
         Args args = new ManagerArgs(argsParam);
         Merge command = new Merge(admin, args);
