@@ -21,24 +21,24 @@ import com.kakao.hbase.TestBase;
 import org.apache.hadoop.hbase.client.HTable;
 import org.apache.hadoop.hbase.client.Put;
 
-public class StatTestBase extends TestBase {
-    protected final int loadEntryLength = RegionLoadAdapter.loadEntries.length;
+class StatTestBase extends TestBase {
+    final int loadEntryLength = RegionLoadAdapter.loadEntries.length;
 
-    public StatTestBase(Class c) {
+    StatTestBase(Class c) {
         super(c);
     }
 
-    protected void putData() throws Exception {
-        try (HTable table = (HTable) hConnection.getTable(tableName)) {
+    void putData() throws Exception {
+        try (HTable table = (HTable) connection.getTable(tableName)) {
             Put put;
             put = new Put("0".getBytes());
-            put.add(TEST_TABLE_CF.getBytes(), "c".getBytes(), "0".getBytes());
+            put.addColumn(TEST_TABLE_CF.getBytes(), "c".getBytes(), "0".getBytes());
             table.put(put);
             put = new Put("a".getBytes());
-            put.add(TEST_TABLE_CF.getBytes(), "c".getBytes(), "a".getBytes());
+            put.addColumn(TEST_TABLE_CF.getBytes(), "c".getBytes(), "a".getBytes());
             table.put(put);
             put = new Put("b".getBytes());
-            put.add(TEST_TABLE_CF.getBytes(), "c".getBytes(), "b".getBytes());
+            put.addColumn(TEST_TABLE_CF.getBytes(), "c".getBytes(), "b".getBytes());
             table.put(put);
         }
     }
