@@ -18,8 +18,9 @@ package com.kakao.hbase.stat;
 
 import com.kakao.hbase.common.LoadEntry;
 import com.kakao.hbase.stat.load.Level;
-import org.apache.hadoop.hbase.HRegionInfo;
 import org.apache.hadoop.hbase.ServerName;
+import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.RegionInfo;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -39,10 +40,10 @@ public class TableStatRsTest extends StatTestBase {
 
         List<ServerName> serverNameList = getServerNameList();
         assertEquals(RS_COUNT, serverNameList.size());
-        String[] args = {"zookeeper", tableName, "--rs", "--interval=0"};
+        String[] args = {"zookeeper", tableName.getNameAsString(), "--rs", "--interval=0"};
         TableStat command = new TableStat(admin, new StatArgs(args));
 
-        ArrayList<HRegionInfo> regionInfoList;
+        ArrayList<RegionInfo> regionInfoList;
 
         // move regions to second RS for checking server index
         regionInfoList = getRegionInfoList(tableName);
@@ -100,11 +101,11 @@ public class TableStatRsTest extends StatTestBase {
         if (miniCluster) {
             args = new String[]{"zookeeper", "--rs", "--interval=0"};
         } else {
-            args = new String[]{"zookeeper", tableName, "--rs", "--interval=0"};
+            args = new String[]{"zookeeper", tableName.getNameAsString(), "--rs", "--interval=0"};
         }
         TableStat command = new TableStat(admin, new StatArgs(args));
 
-        ArrayList<HRegionInfo> regionInfoList;
+        ArrayList<RegionInfo> regionInfoList;
 
         // move regions to second RS for checking server index
         regionInfoList = getRegionInfoList(tableName);
@@ -157,14 +158,14 @@ public class TableStatRsTest extends StatTestBase {
         if (miniCluster) {
             splitTable("a".getBytes());
 
-            String tableName2 = createAdditionalTable(tableName + "2");
+            TableName tableName2 = createAdditionalTable(tableName + "2");
 
             List<ServerName> serverNameList = getServerNameList();
             assertEquals(RS_COUNT, serverNameList.size());
             String[] args = {"zookeeper", "--rs", "--interval=0"};
             TableStat command = new TableStat(admin, new StatArgs(args));
 
-            ArrayList<HRegionInfo> regionInfoList;
+            ArrayList<RegionInfo> regionInfoList;
 
             // move regions to second RS for checking server index
             regionInfoList = getRegionInfoList(tableName);
